% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{magp2d_rmse}
\alias{magp2d_rmse}
\title{Calculate root-mean-squared prediction error}
\usage{
magp2d_rmse(pred, actual)
}
\arguments{
\item{pred}{Numeric vector of predictions.}

\item{actual}{Numeric vector of observed values with the same length as
\code{pred}.}
}
\value{
A single nonnegative numeric value.
}
\description{
Calculates the root-mean-squared error between predicted and
observed values.
}
