% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-optimization.R
\name{print.magp_bayes_opt}
\alias{print.magp_bayes_opt}
\title{Print a MaGP Bayesian optimization result}
\usage{
\method{print}{magp_bayes_opt}(x, ...)
}
\arguments{
\item{x}{A \code{magp_bayes_opt} object returned by
\code{\link[=magp_bayes_optimize]{magp_bayes_optimize()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a MaGP Bayesian optimization result
}
