% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-design-complete.R
\name{print.magp_initial_design}
\alias{print.magp_initial_design}
\title{Print a quantitative-sequence initial design}
\usage{
\method{print}{magp_initial_design}(x, ...)
}
\arguments{
\item{x}{A \code{magp_initial_design} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a quantitative-sequence initial design
}
