% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition.R
\name{print.magp_next_point}
\alias{print.magp_next_point}
\title{Print a MaGP acquisition-search result}
\usage{
\method{print}{magp_next_point}(x, ...)
}
\arguments{
\item{x}{A \code{magp_next_point} object returned by \code{\link[=magp_next_point]{magp_next_point()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a MaGP acquisition-search result
}
