% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-design-quantitative.R
\name{print.magp_quantitative_design}
\alias{print.magp_quantitative_design}
\title{Print a quantitative initial design}
\usage{
\method{print}{magp_quantitative_design}(x, ...)
}
\arguments{
\item{x}{A \code{magp_quantitative_design} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a quantitative initial design
}
