% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-design-sequence.R
\name{print.magp_sequence_design}
\alias{print.magp_sequence_design}
\title{Print a sequence initial design}
\usage{
\method{print}{magp_sequence_design}(x, ...)
}
\arguments{
\item{x}{A fitted \code{magp_sequence_design} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a sequence initial design
}
